implement checks on the backend

+++ backend/app/Http/Controllers/StudentController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Student;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Resources\StudentResource;
 use App\Http\Requests\StoreStudentRequest;
 use App\Http\Requests\UpdateStudentRequest;
@@ -12,6 +13,8 @@ class StudentController extends Controller
 {
     public function index(Request $request)
     {
+        Gate::authorize('student_list');
+
         $students = Student::search($request);
 
         return StudentResource::collection(
@@ -21,6 +24,8 @@ class StudentController extends Controller
 
     public function store(StoreStudentRequest $request)
     {
+        Gate::authorize('student_create');
+
         Student::create($request->validated());
 
         return response()->json([
@@ -30,6 +35,8 @@ class StudentController extends Controller
 
     public function update(UpdateStudentRequest $request, Student $student)
     {
+        Gate::authorize('student_edit');
+
         $student->update($request->validated());
 
         return response()->json([
@@ -39,11 +46,15 @@ class StudentController extends Controller
 
     public function show(Student $student)
     {
+        Gate::authorize('student_show');
+
         return StudentResource::make($student);
     }
 
     public function destroy(Student $student)
     {
+        Gate::authorize('student_delete');
+
         $student->delete();
 
         return response()->json([
